Split IdP SQL providers
