+++ plugins/idp-basic/src/main/java/org/apache/gravitino/idp/basic/storage/relational/mapper/IdpUserMetaSQLProviderFactory.java
@@ -23,6 +23,8 @@ import com.google.common.collect.ImmutableMap;
 import java.util.List;
 import java.util.Map;
 import org.apache.gravitino.idp.basic.storage.relational.mapper.provider.base.IdpUserMetaBaseSQLProvider;
+import org.apache.gravitino.idp.basic.storage.relational.mapper.provider.mysql.IdpUserMetaMySQLProvider;
+import org.apache.gravitino.idp.basic.storage.relational.mapper.provider.postgresql.IdpUserMetaPostgreSQLProvider;
 import org.apache.gravitino.idp.basic.storage.relational.po.IdpUserPO;
 import org.apache.gravitino.storage.relational.JDBCBackend.JDBCBackendType;
 import org.apache.gravitino.storage.relational.session.SqlSessionFactoryHelper;
@@ -124,32 +126,6 @@ public class IdpUserMetaSQLProviderFactory {
     @Override
     protected String currentTimeMillisExpression() {
       return "DATEDIFF('MILLISECOND', TIMESTAMP '1970-01-01 00:00:00', CURRENT_TIMESTAMP())";
-    }
-  }
-
-  static class IdpUserMetaMySQLProvider extends IdpUserMetaBaseSQLProvider {
-
-    @Override
-    protected String currentTimeMillisExpression() {
-      return "(UNIX_TIMESTAMP() * 1000.0)";
-    }
-  }
-
-  static class IdpUserMetaPostgreSQLProvider extends IdpUserMetaBaseSQLProvider {
-
-    @Override
-    protected String currentTimeMillisExpression() {
-      return "CAST(EXTRACT(EPOCH FROM CURRENT_TIMESTAMP) * 1000 AS BIGINT)";
-    }
-
-    @Override
-    public String deleteIdpUserMetasByLegacyTimeline(
-        @Param("legacyTimeline") Long legacyTimeline, @Param("limit") int limit) {
-      return "DELETE FROM "
-          + IdpUserMetaMapper.IDP_USER_TABLE_NAME
-          + " WHERE user_id IN (SELECT user_id FROM "
-          + IdpUserMetaMapper.IDP_USER_TABLE_NAME
-          + " WHERE deleted_at > 0 AND deleted_at < #{legacyTimeline} LIMIT #{limit})";
     }
   }
 }
